refactor(extension-common): タブのバッジ更新の配線を tab-badge へ移す
link-verification と同様に、tab-badge も自前のモジュールを持ちながら Service
Worker 側の配線が setupBackground に残っていた。tab-badge/background.ts に
setupTabBadge(countCredentials) として切り出す。setupBackground は 186 行に
なり、残るのはサイドパネル / サイドバー、インストール時の初期化、フレーム CAS、
Basic Auth の 4 話題になった。

onInstalled は分けていない。インストール直後のバッジ更新は Content Script の
注入完了を待つ必要があり (countCredentials はタブへメッセージを送る)、
リスナーを 2 本に割ると両者の順序が保証されなくなる。setupTabBadge が
requestTabBadgeUpdate を返し、呼び出し側が注入の後に呼ぶ形にした。

onRemoved を分けたときと判断が分かれるが、あちらは互いに素な状態を破棄する
だけで依存がなく、こちらには依存がある。
